Refactor startup migrations.
