+++ lsif/src/server/server.ts
@@ -1,12 +1,9 @@
 import * as constants from '../shared/constants'
-import * as fs from 'mz/fs'
 import * as metrics from './metrics'
 import * as path from 'path'
 import * as settings from './settings'
 import express from 'express'
 import promClient from 'prom-client'
-import { createClient } from 'redis'
-import { promisify } from 'util'
 import { Backend } from './backend/backend'
 import { createLogger } from '../shared/logging'
 import { createLsifRouter } from './routes/lsif'
@@ -14,7 +11,7 @@ import { createMetaRouter } from './routes/meta'
 import { createPostgresConnection } from '../shared/database/postgres'
 import { createTracer } from '../shared/tracing'
 import { createUploadRouter } from './routes/uploads'
-import { dbFilename, ensureDirectory, idFromFilename } from '../shared/paths'
+import { ensureDirectory } from '../shared/paths'
 import { default as tracingMiddleware } from 'express-opentracing'
 import { errorHandler } from './middleware/errors'
 import { logger as loggingMiddleware } from 'express-winston'
@@ -26,6 +23,7 @@ import { waitForConfiguration } from '../shared/config/config'
 import { DumpManager } from '../shared/store/dumps'
 import { DependencyManager } from '../shared/store/dependencies'
 import { SRC_FRONTEND_INTERNAL } from '../shared/config/settings'
+import { migrate } from './startup-migrations/migration'
 
 /**
  * Runs the HTTP server that accepts LSIF dump uploads and responds to LSIF requests.
@@ -60,9 +58,15 @@ async function main(logger: Logger): Promise<void> {
     const dependencyManager = new DependencyManager(connection)
     const backend = new Backend(settings.STORAGE_ROOT, dumpManager, dependencyManager, SRC_FRONTEND_INTERNAL)
 
-    // Temporary migration
-    await migrateFilenames() // TODO - remove after 3.15
-    await clearOldRedisData(logger) // TODO - remove after 3.15
+    // Run any app-level migrations. These migrations usually exist only
+    // for a two-minor-version period in which we clean up old data and
+    // fix outdated assumptions.
+    //
+    // These block the process from starting up until completion. Also
+    // note that if the cleanup is handling an assumption from the last
+    // minor version, there may be instances of that version running
+    // after this migration step completes.
+    await migrate({ logger }, connection, dumpManager)
 
     // Start background tasks
     startTasks(connection, dumpManager, uploadManager, logger)
@@ -93,59 +97,6 @@ async function main(logger: Logger): Promise<void> {
     app.use(errorHandler(logger))
 
     app.listen(settings.HTTP_PORT, () => logger.debug('LSIF API server listening on', { port: settings.HTTP_PORT }))
-}
-
-/**
- * If it hasn't been done already, migrate from the old pre-3.13 filename format
- * `$ID-$REPO@$COMMIT.lsif.db` to the new format `$ID.lsif.db`.
- */
-async function migrateFilenames(): Promise<void> {
-    const doneFile = path.join(settings.STORAGE_ROOT, 'id-only-based-filenames')
-    if (await fs.exists(doneFile)) {
-        // Already migrated.
-        return
-    }
-
-    for (const basename of await fs.readdir(path.join(settings.STORAGE_ROOT, constants.DBS_DIR))) {
-        const id = idFromFilename(basename)
-        if (!id) {
-            continue
-        }
-
-        await fs.rename(
-            path.join(settings.STORAGE_ROOT, constants.DBS_DIR, basename),
-            dbFilename(settings.STORAGE_ROOT, id)
-        )
-    }
-
-    // Create an empty done file to record that all files have been renamed.
-    await fs.close(await fs.open(doneFile, 'w'))
-}
-
-/**
- * Remove all old LSIF data from redis.
- */
-async function clearOldRedisData(logger: Logger): Promise<void> {
-    const script = `
-        for i, key in ipairs(redis.call('keys', 'lsif:*')) do
-            redis.call('del', key)
-        end
-
-        for i, key in ipairs(redis.call('keys', 'bull:*')) do
-            redis.call('del', key)
-        end
-    `
-
-    const url = process.env.REDIS_STORE_ENDPOINT || process.env.REDIS_ENDPOINT || 'redis-store:6379'
-    const urlWithProtocol = url.includes('//') ? url : `redis://${url}`
-
-    try {
-        const client = createClient(urlWithProtocol)
-        const evalAsync: (script: string, numArgs: number) => Promise<void> = promisify(client.eval).bind(client)
-        await evalAsync(script, 0)
-    } catch (err) {
-        logger.warning('Failed to clean old LSIF data from redis-store', { error: err })
-    }
 }
 
 // Initialize logger
